move set_layer_norm_eps_recursively from bert.py to layer.py
