+++ axlearn/common/bert.py
@@ -35,7 +35,7 @@ from axlearn.common.attention import (
     scaled_hidden_dim,
 )
 from axlearn.common.base_layer import BaseLayer, ParameterSpec
-from axlearn.common.config import REQUIRED, ConfigBase, InstantiableConfig, Required, config_class
+from axlearn.common.config import REQUIRED, InstantiableConfig, Required, config_class
 from axlearn.common.embedding import TransformerTextEmbeddings
 from axlearn.common.encoder import Encoder, EncoderModel
 from axlearn.common.layers import (
@@ -46,6 +46,7 @@ from axlearn.common.layers import (
     Linear,
     RedirectToSharedModule,
     get_activation_fn,
+    set_layer_norm_eps_recursively,
 )
 from axlearn.common.loss import mean_squared_error
 from axlearn.common.module import Module, child_context
@@ -341,28 +342,6 @@ class BertModel(EncoderModel):
             )
         )
         return cfg
-
-
-def set_layer_norm_eps_recursively(cfg: ConfigBase, eps: float, set_only_if_none: bool = False):
-    """Sets LayerNorm.Config.eps recursively.
-
-    Args:
-        cfg: The root config under which to look for LayerNorm.Config.
-        eps: The target value.
-        set_only_if_none: Override LayerNorm.Config.eps to `eps` only if the original is None.
-    """
-
-    def is_layer_norm_config(cfg):
-        return isinstance(cfg, LayerNorm.Config)
-
-    def visit_fn(_, value):
-        if is_layer_norm_config(value) and (not set_only_if_none or value.eps is None):
-            value.eps = eps
-
-    def enter_fn(_, value, default_kv):
-        return None if is_layer_norm_config(value) else default_kv
-
-    cfg.visit(visit_fn=visit_fn, enter_fn=enter_fn)
 
 
 def bert_embedding_config(
